rename clusters files
