+++ webui/react/src/pages/Cluster/ClusterQueuedChart.module.scss
